test: reorganize suite by module under test and remove duplicates
Group tests by the unit they cover instead of by bug-fix batch:
- add tests/conftest.py (autouse KONFAI env defaults, write_config and
  image_attributes fixtures), replacing 12 copy-pasted module-level
  os.environ.setdefault blocks that violated the monkeypatch rule
- fold the audit/fixes grab-bags and single-test files into per-module
  files: config, patching, data_manager, transform, augmentation,
  network, blocks, schedulers, models, dataset, measure, main_cli
- merge 5 duplicated tests into their surviving twin, keeping every
  assertion (config missing-file/invalid-yaml/invalid-literal errors,
  Data._split tail dedup)

38 unit files instead of 60; 343 -> 338 collected tests (the 5 exact
duplicates); per-module coverage is byte-identical before/after.
