+++ tests/unit/test_yaml_model_equivalence.py
@@ -22,16 +22,11 @@ must produce a graph with the same parameter count and forward behaviour as the
 hand-written ``konfai.models.segmentation.UNet`` configured identically.
 """
 
-import os
+from pathlib import Path
 
-os.environ.setdefault("KONFAI_config_file", "/tmp/konfai-none.yml")
-os.environ.setdefault("KONFAI_CONFIG_MODE", "Done")
-
-from pathlib import Path  # noqa: E402
-
-import torch  # noqa: E402
-from konfai.network.blocks import BlockConfig  # noqa: E402
-from konfai.utils.model_builder import build_model_from_yaml  # noqa: E402
+import torch
+from konfai.network.blocks import BlockConfig
+from konfai.utils.model_builder import build_model_from_yaml
 
 UNET_YML = Path(__file__).resolve().parents[2] / "examples" / "Segmentation" / "UNet.yml"
 
@@ -53,9 +48,7 @@ def test_example_unet_yaml_matches_python_unet_param_count():
     from konfai.models.segmentation.UNet import UNet
 
     yaml_net = _build_yaml_unet()
-    block_config = BlockConfig(
-        kernel_size=3, stride=1, padding=1, bias=True, activation="ReLU", norm_mode="NONE"
-    )
+    block_config = BlockConfig(kernel_size=3, stride=1, padding=1, bias=True, activation="ReLU", norm_mode="NONE")
     python_net = UNet(
         dim=2,
         channels=[1, 32, 64, 128, 256],
